+++ nodes/SystemMonitor/node/SystemMonitorProcess.cpp
@@ -1,0 +1,58 @@
+#include "SystemMonitorProcess.h"
+
+using namespace eros;
+namespace eros_nodes::SystemMonitor {
+SystemMonitorProcess::~SystemMonitorProcess() {
+    delete header_window;
+}
+Diagnostic::DiagnosticDefinition SystemMonitorProcess::finish_initialization() {
+    Diagnostic::DiagnosticDefinition diag = diagnostic_helper.get_root_diagnostic();
+    diag = diagnostic_helper.update_diagnostic(Diagnostic::DiagnosticType::SOFTWARE,
+                                               Level::Type::INFO,
+                                               Diagnostic::Message::NOERROR,
+                                               "Finished Initialization.");
+    return diag;
+}
+void SystemMonitorProcess::reset() {
+}
+Diagnostic::DiagnosticDefinition SystemMonitorProcess::update(double t_dt, double t_ros_time) {
+    Diagnostic::DiagnosticDefinition diag = base_update(t_dt, t_ros_time);
+    int key_pressed = getch();
+    logger->log_warn("Key: " + std::to_string(key_pressed));
+    if ((key_pressed == KEY_q) || (key_pressed == KEY_Q)) {
+        kill_me = true;
+    }
+
+    header_window->update();
+    flushinp();
+
+    return diag;
+}
+std::vector<Diagnostic::DiagnosticDefinition> SystemMonitorProcess::new_commandmsg(
+    eros::command msg) {
+    (void)msg;  // Not currently used.
+    std::vector<Diagnostic::DiagnosticDefinition> diag_list;
+    Diagnostic::DiagnosticDefinition diag = diagnostic_helper.get_root_diagnostic();
+    diag = diagnostic_helper.update_diagnostic(Diagnostic::DiagnosticType::SOFTWARE,
+                                               Level::Type::INFO,
+                                               Diagnostic::Message::NOERROR,
+                                               "Processed Command.");
+    diag_list.push_back(diag);
+    return diag_list;
+}
+std::vector<Diagnostic::DiagnosticDefinition> SystemMonitorProcess::check_programvariables() {
+    std::vector<Diagnostic::DiagnosticDefinition> diag_list;
+    Diagnostic::DiagnosticDefinition diag = diagnostic_helper.get_root_diagnostic();
+    diag = diagnostic_helper.update_diagnostic(Diagnostic::DiagnosticType::SOFTWARE,
+                                               Level::Type::INFO,
+                                               Diagnostic::Message::NOERROR,
+                                               "Program Variable Check Passed.");
+    diag_list.push_back(diag);
+    return diag_list;
+}
+bool SystemMonitorProcess::initialize_windows() {
+    timeout(0);
+    header_window = new HeaderWindow(logger, mainwindow_height, mainwindow_width);
+    return true;
+}
+}  // namespace eros_nodes::SystemMonitor
