+++ nodes/SystemMonitor/src/DeviceWindow/DeviceWindow.h
@@ -1,0 +1,11 @@
+#pragma once
+#include <BaseWindow.h>
+class DeviceWindow : public BaseWindow
+{
+   public:
+    DeviceWindow();
+    virtual ~DeviceWindow();
+    bool update() override;
+
+   private:
+}
