+++ nodes/SystemMonitor/src/HeaderWindow/CMakeLists.txt
@@ -1,0 +1,2 @@
+add_library(headerWindow HeaderWindow.cpp)
+target_link_libraries(headerWindow ${catkin_LIBRARIES} ${CURSES_LIBRARIES} baseWindow)
